==> 5_Dhruv_Kruthi_RoBERTa_Clean (1).ipynb <==
# Computing RoBERTa embeddings for Cleaned headlines

import pandas as pd

norm_df = pd.read_csv('new_normalized_headlines.csv')

norm_df.head()

from sklearn.model_selection import train_test_split
# Split the data into train and test samples
X_train, X_test, y_train, y_test = train_test_split(norm_df['cleaned'], norm_df['news'], test_size=0.2, random_state=42, stratify=norm_df['news'])

# 3. Convert labels to binary (1 = FoxNews, 0 = NBC)
y_train = y_train.apply(lambda x: 1 if x == 'Fox News' else 0)
y_test = y_test.apply(lambda x: 1 if x == 'Fox News' else 0)

from transformers import RobertaTokenizer, RobertaModel
import torch

# Load pre-trained RoBERTa model and tokenizer
tokenizer = RobertaTokenizer.from_pretrained('roberta-base')
model = RobertaModel.from_pretrained('roberta-base')

model.eval()

device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
print(f"Using device: {device}")
model = model.to(device)

def get_embeddings(texts, tokenizer, model, max_len=128):
    embeddings = []

    for text in texts:
        # Tokenize
        inputs = tokenizer(
            text,
            add_special_tokens=True,
            max_length=max_len,
            padding='max_length',
            truncation=True,
            return_tensors='pt'
        )

        # Move to the same device as the model
        inputs = {k: v.to(model.device) for k, v in inputs.items()}

        with torch.no_grad():
            outputs = model(**inputs)
            # Extract <s> token embedding, move to CPU, convert to NumPy
            cls_embedding = outputs.last_hidden_state[:, 0, :].squeeze().cpu().numpy()
            embeddings.append(cls_embedding)

    return embeddings

# Get train and test embeddings (this may take awhile)
train_embeddings = get_embeddings(X_train, tokenizer, model)
test_embeddings = get_embeddings(X_test, tokenizer, model)

print(len(train_embeddings))

import numpy as np
print(np.array(train_embeddings).shape)

print(np.array(test_embeddings).shape)

import numpy as np

# Convert list of embeddings to a NumPy array
train_embeddings_array = np.array(train_embeddings)
test_embeddings_array = np.array(test_embeddings)

train_embeddings_array.shape

np.save('RoBERTa_train_clean_embeddings.npy', train_embeddings_array)
np.save('RoBERTa_test_clean_embeddings.npy', test_embeddings_array)

y_train = np.array(y_train)
y_test = np.array(y_test)
np.save('y_train_roberta_clean.npy', y_train)
np.save('y_test_roberta_clean.npy', y_test)
